+++ packages/components/data-grid/src/data-grid-table.tsx
@@ -1,20 +1,28 @@
 "use client";
-import { Table } from "@jamsrui/table";
 
-import { useDataGridContext } from "./data-grid-context";
+import { Table } from "@jamsrui/table";
+import { DataGridBody } from "./data-grid-body";
+import { DataGridHeader } from "./data-grid-header";
+import { cn } from "@jamsrui/utils";
 
 export const DataGridTable = (props: DataGridTable.Props) => {
-  const { children } = props;
-  const { rootProps } = useDataGridContext();
+  const {
+    children = (
+      <>
+        <DataGridHeader />
+        <DataGridBody />
+      </>
+    ),
+    className,
+    ...restProps
+  } = props;
   return (
-    <Table className="w-full table-fixed" {...rootProps}>
+    <Table className={cn("w-full table-fixed", className)} {...restProps}>
       {children}
     </Table>
   );
 };
 
 export namespace DataGridTable {
-  export interface Props {
-    children: React.ReactNode;
-  }
+  export interface Props extends Table.Props {}
 }