+++ packages/components/menu/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsrui/menu
 
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsrui/icons@0.0.16
+
 ## 0.0.21
 
 ### Patch Changes