+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsrui/icons
 
+## 0.0.16
+
+### Patch Changes
+
+- fix: data-grid
+
 ## 0.0.15
 
 ### Patch Changes